requirements.txt
Django==5.0.6
django-browser-reload==1.12.1
django-tailwind==3.8.0
pillow==10.3.0
nada-dsl==0.2.1
py-nillion-client==0.2.1
python-dotenv==1.0.0
django-celery-beat==2.6.0